+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- `opentelemetry-instrumentation-django`: Fix `_DjangoMiddleware` to avoid hitting a host check ([#3651](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3651))
+
 ### Fixed
 
 - `opentelemetry-instrumentation-botocore`: migrate off the deprecated events API to use the logs API